Connect plan functions to best first search

+++ Planner/src/search/BestFirstSearch.java
@@ -70,7 +70,7 @@ public class BestFirstSearch
     private boolean filterPreviousState (final SearchState child)
     {
 	final ProblemState p = child.getProblemState ();
-	final int cost = child.getCost ();
+	final double cost = child.getCost ();
 	for (final SearchState s : closed)
 	{
 	    if (p == s.getProblemState ())
